refactor: move httpHelper to a folder
issue: #6
