+++ src/presentation/controllers/login/login.spec.ts
@@ -1,6 +1,6 @@
 import { HttpRequest,Authentication,Validation } from './login-protocols'
 import { MissingParamError } from '../../errors'
-import { badRequest, ok, serverError, unaunthorized } from '../../helpers/http-helper'
+import { badRequest, ok, serverError, unaunthorized } from '../../helpers/http/http-helper'
 import { LoginController } from './login'
 
 const MakeAuthentication = (): Authentication => {
