+++ src/presentation/controllers/login/login.ts
@@ -1,5 +1,5 @@
 import { Controller, HttpRequest, HttpResponse,Authentication,Validation } from './login-protocols'
-import { badRequest, ok, serverError, unaunthorized } from '../../helpers/http-helper'
+import { badRequest, ok, serverError, unaunthorized } from '../../helpers/http/http-helper'
 
 export class LoginController implements Controller {
   constructor (
