+++ src/presentation/controllers/signup/signup.spec.ts
@@ -2,7 +2,7 @@ import { SignUpController } from './signup'
 import { MissingParamError,ServerError } from '../../errors'
 import { AccountModel,AddAccount,AddAccountModel,Validation } from './signup-protocols'
 import { HttpRequest } from '../../protocols'
-import { badRequest, ok, serverError } from '../../helpers/http-helper'
+import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 
 const MakeFakeAccount = (): AccountModel => ({
   id: 'valid_id',
