+++ src/presentation/controllers/signup/signup.ts
@@ -1,5 +1,5 @@
 import { HttpResponse,HttpRequest,Controller, AddAccount,Validation } from './signup-protocols'
-import { badRequest, ok, serverError } from '../../helpers/http-helper'
+import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 
 export class SignUpController implements Controller {
   constructor (
