+++ src/presentation/helpers/http/http-helper.ts
@@ -1,5 +1,5 @@
-import { ServerError, UnaunthorizedError } from '../errors'
-import { HttpResponse } from '../protocols/http'
+import { ServerError, UnaunthorizedError } from '../../errors'
+import { HttpResponse } from '../../protocols/http'
 
 export const badRequest = (error: Error): HttpResponse => ({
   statusCode: 400,
